re-working the consumer client website into a more structured situation
